Activiti/Activiti#4285: Move PublicKeyValidationCheck to common-security project
